% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_english_to_new_sat_writing.R
\name{act_english_to_new_sat_writing}
\alias{act_english_to_new_sat_writing}
\title{Recode ACT Enlish/Writing test to new SAT writing and language section}
\usage{
act_english_to_new_sat_writing(x)
}
\arguments{
\item{x}{numeric vector of ACT Enlish/Writing test values}
}
\value{
integer vector of new SAT writing and language section values
}
\description{
Function to recode ACT Enlish/Writing test to new SAT writing and language section
}
