% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_to_new_sat.R
\name{act_to_new_sat}
\alias{act_to_new_sat}
\title{Recode ACT to new SAT (1600 scale)}
\usage{
act_to_new_sat(x)
}
\arguments{
\item{x}{numeric vector of ACT values}
}
\value{
integer vector of new SAT values (1600 scale)
}
\description{
Function to recode ACT to new SAT (1600 scale)
}
