% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_math_to_old_sat_math.R
\name{new_sat_math_to_old_sat_math}
\alias{new_sat_math_to_old_sat_math}
\title{Recode new SAT math section to old SAT math section}
\usage{
new_sat_math_to_old_sat_math(x)
}
\arguments{
\item{x}{numeric vector of new SAT math section values}
}
\value{
integer vector of old SAT math section values
}
\description{
Function to recode new SAT math section to old SAT math section
}
