% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_reading_to_old_sat_reading.R
\name{new_sat_reading_to_old_sat_reading}
\alias{new_sat_reading_to_old_sat_reading}
\title{Recode new SAT reading test to old SAT critical reading section}
\usage{
new_sat_reading_to_old_sat_reading(x)
}
\arguments{
\item{x}{numeric vector of new SAT reading test values}
}
\value{
integer vector of old SAT critical reading section values
}
\description{
Function to recode new SAT reading test to old SAT critical reading section
}
