% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_to_act.R
\name{new_sat_to_act}
\alias{new_sat_to_act}
\title{Recode new SAT (1600 scale) to ACT}
\usage{
new_sat_to_act(x)
}
\arguments{
\item{x}{numeric vector of new SAT values (1600 scale)}
}
\value{
integer vector of ACT values
}
\description{
Function to recode new SAT (1600 scale) to ACT.
Missing and unmatched values are coded as missing.
}
\examples{
require(dplyr)

new_sat_vals <- round(sample(560:1600, 10), -1)
act_vals <- new_sat_to_act(new_sat_vals)
new_act_vals <- data.frame(new_sat_vals, act_vals)
print(new_act_vals)

new_act_vals \%>\%
    mutate(act_vals_piped = new_sat_to_act(new_sat_vals))
}
