% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_to_old_sat_1600.R
\name{new_sat_to_old_sat_1600}
\alias{new_sat_to_old_sat_1600}
\title{Recode new SAT to old SAT (1600 scale)}
\usage{
new_sat_to_old_sat_1600(x)
}
\arguments{
\item{x}{numeric vector of new SAT values on 1600 scale}
}
\value{
integer vector of old SAT values on 1600 scale
}
\description{
Function to recode new SAT to old SAT (1600 scale)
}
