% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_to_old_sat_2400.R
\name{new_sat_to_old_sat_2400}
\alias{new_sat_to_old_sat_2400}
\title{Recode new SAT to old SAT (2400 scale)}
\usage{
new_sat_to_old_sat_2400(x)
}
\arguments{
\item{x}{numeric vector of new SAT values on 2400 scale}
}
\value{
integer vector of old SAT values on 2400 scale
}
\description{
Function to recode new SAT to old SAT (2400 Scale)
}
