% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_writing_to_act_english.R
\name{new_sat_writing_to_act_english}
\alias{new_sat_writing_to_act_english}
\title{Recode new SAT writing and language section to ACT Enlish/Writing test}
\usage{
new_sat_writing_to_act_english(x)
}
\arguments{
\item{x}{numeric vector of new SAT writing and language section values}
}
\value{
integer vector of ACT Enlish/Writing test values
}
\description{
Function to recode new SAT writing and language section to ACT Enlish/Writing test
}
