% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_writing_to_old_sat_writing.R
\name{new_sat_writing_to_old_sat_writing}
\alias{new_sat_writing_to_old_sat_writing}
\title{Recode new SAT writing and language test to old SAT writing section}
\usage{
new_sat_writing_to_old_sat_writing(x)
}
\arguments{
\item{x}{numeric vector of new SAT writing and language values}
}
\value{
integer vector of old SAT writing section values
}
\description{
Function to recode new SAT writing and language test to old SAT writing section
}
