% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_math_to_new_sat_math_test.R
\name{old_sat_math_to_new_sat_math_test}
\alias{old_sat_math_to_new_sat_math_test}
\title{Recode old SAT math section to new SAT math test (10-40 scale)}
\usage{
old_sat_math_to_new_sat_math_test(x)
}
\arguments{
\item{x}{numeric vector of old SAT math section values}
}
\value{
integer vector of new SAT math test values (10-40 scale)
}
\description{
Function to recode old SAT math section to new SAT math test (10-40 scale)
}
