% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_reading_to_new_sat_reading.R
\name{old_sat_reading_to_new_sat_reading}
\alias{old_sat_reading_to_new_sat_reading}
\title{Recode old SAT critical reading test to new SAT reading section}
\usage{
old_sat_reading_to_new_sat_reading(x)
}
\arguments{
\item{x}{numeric vector of old SAT critical reading test values}
}
\value{
integer vector of new SAT reading section values
}
\description{
Function to recode old SAT critical reading test to new SAT reading section
}
\examples{
require(dplyr)

old_sat_vals <- round(sample(200:800, 10), -1)
new_sat_vals <- old_sat_reading_to_new_sat_reading(old_sat_vals)
old_new_vals <- data.frame(old_sat_vals, new_sat_vals)
print(old_new_vals)

old_new_vals \%>\%
    mutate(new_sat_vals_piped = old_sat_reading_to_new_sat_reading(old_sat_vals))
}
