% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_to_new_sat_2400.R
\name{old_sat_to_new_sat_2400}
\alias{old_sat_to_new_sat_2400}
\title{Recode old SAT to new SAT (2400 scale)}
\usage{
old_sat_to_new_sat_2400(x)
}
\arguments{
\item{x}{numeric vector of old SAT values on 2400 scale}
}
\value{
integer vector of new SAT values on 2400 scale
}
\description{
Function to recode old SAT to new SAT (2400 Scale)
}
