% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_writing_to_new_sat_writing.R
\name{old_sat_writing_to_new_sat_writing}
\alias{old_sat_writing_to_new_sat_writing}
\title{Recode old SAT writing test to new SAT writing and language  section}
\usage{
old_sat_writing_to_new_sat_writing(x)
}
\arguments{
\item{x}{numeric vector of old SAT writing values}
}
\value{
integer vector of new SAT writing and language section values
}
\description{
Function to recode old SAT writing test to new SAT writing and language section
}
