% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_writingreading_to_new_ebrw.R
\name{old_sat_writingreading_to_new_ebrw}
\alias{old_sat_writingreading_to_new_ebrw}
\title{Recode old SAT writing plus critical reading sections to new SAT EBRW section}
\usage{
old_sat_writingreading_to_new_ebrw(x)
}
\arguments{
\item{x}{numeric vector of old SAT writing plus critical reading section values}
}
\value{
integer vector of new SAT EBRW section values
}
\description{
Function to recode old SAT writing plus critical reading sections
to new SAT Evidence-Based Reading and Writing Section test.
Missing and unmatched values are coded as missing.
}
\examples{
require(dplyr)

old_sat_vals <- round(sample(400:1600, 10), -1)
new_sat_vals <- old_sat_writingreading_to_new_ebrw(old_sat_vals)
old_new_vals <- data.frame(old_sat_vals, new_sat_vals)
print(old_new_vals)

old_new_vals \%>\%
    mutate(new_sat_vals_piped = old_sat_writingreading_to_new_ebrw(old_sat_vals))
}
