% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_to_new_sat.R
\name{act_to_new_sat}
\alias{act_to_new_sat}
\title{Recode ACT to new SAT (1600 scale)}
\usage{
act_to_new_sat(x)
}
\arguments{
\item{x}{numeric vector of ACT values}
}
\value{
integer vector of new SAT values (1600 scale)
}
\description{
Function to recode ACT to new SAT (1600 scale).
Missing and unmatched values are coded as missing.
}
\examples{
require(dplyr)

act_vals <- sample(11:36, 10)
new_sat_vals <- act_to_new_sat(act_vals)
act_new_vals <- data.frame(act_vals, new_sat_vals)
print(act_new_vals)

act_new_vals \%>\%
    mutate(new_sat_vals_piped = act_to_new_sat(act_vals))
}
