% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_ebrw_to_old_writingreading.R
\name{new_sat_ebrw_to_old_writingreading}
\alias{new_sat_ebrw_to_old_writingreading}
\title{Recode new SAT EBRW section to old SAT writing plus critical reading sections}
\usage{
new_sat_ebrw_to_old_writingreading(x)
}
\arguments{
\item{x}{numeric vector of new SAT EBRW section values}
}
\value{
integer vector of old SAT writing plus critical reading section values
}
\description{
Function to recode new SAT Evidence-Based Reading and Writing Section test
to old SAT writing plus critical reading sections
}
\examples{
require(dplyr)

new_sat_vals <- round(sample(200:800, 10), -1)
old_sat_vals <- new_sat_ebrw_to_old_writingreading(new_sat_vals)
new_old_vals <- data.frame(new_sat_vals, old_sat_vals)
print(new_old_vals)

new_old_vals \%>\%
    mutate(old_sat_vals_piped = new_sat_ebrw_to_old_writingreading(new_sat_vals))
}
