% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_math_to_old_sat_math.R
\name{new_sat_math_to_old_sat_math}
\alias{new_sat_math_to_old_sat_math}
\title{Recode new SAT math section to old SAT math section}
\usage{
new_sat_math_to_old_sat_math(x)
}
\arguments{
\item{x}{numeric vector of new SAT math section values}
}
\value{
integer vector of old SAT math section values
}
\description{
Function to recode new SAT math section to old SAT math section
}
\examples{
require(dplyr)

new_sat_vals <- round(sample(200:800, 10), -1)
old_sat_vals <- new_sat_math_to_old_sat_math(new_sat_vals)
new_old_vals <- data.frame(new_sat_vals, old_sat_vals)
print(new_old_vals)

new_old_vals \%>\%
    mutate(old_sat_vals_piped = new_sat_math_to_old_sat_math(new_sat_vals))
}
