% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_reading_to_old_sat_reading.R
\name{new_sat_reading_to_old_sat_reading}
\alias{new_sat_reading_to_old_sat_reading}
\title{Recode new SAT reading test to old SAT critical reading section}
\usage{
new_sat_reading_to_old_sat_reading(x)
}
\arguments{
\item{x}{numeric vector of new SAT reading test values}
}
\value{
integer vector of old SAT critical reading section values
}
\description{
Function to recode new SAT reading test to old SAT critical reading section
}
\examples{
require(dplyr)

new_sat_vals <- sample(10:40, 10)
old_sat_vals <- new_sat_reading_to_old_sat_reading(new_sat_vals)
new_old_vals <- data.frame(new_sat_vals, old_sat_vals)
print(new_old_vals)

new_old_vals \%>\%
    mutate(old_sat_vals_piped = new_sat_reading_to_old_sat_reading(new_sat_vals))
}
