% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_to_act.R
\name{new_sat_to_act}
\alias{new_sat_to_act}
\title{Recode new SAT (1600 scale) to ACT}
\usage{
new_sat_to_act(x)
}
\arguments{
\item{x}{numeric vector of new SAT values (1600 scale)}
}
\value{
integer vector of ACT values
}
\description{
Function to recode new SAT (1600 scale) to ACT
}
