% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_writing_to_act_english.R
\name{new_sat_writing_to_act_english}
\alias{new_sat_writing_to_act_english}
\title{Recode new SAT writing and language section to ACT Enlish/Writing test}
\usage{
new_sat_writing_to_act_english(x)
}
\arguments{
\item{x}{numeric vector of new SAT writing and language section values}
}
\value{
integer vector of ACT Enlish/Writing test values
}
\description{
Function to recode new SAT writing and language section to ACT Enlish/Writing test
}
\examples{
require(dplyr)

new_sat_vals <- sample(17:40, 10)
act_vals <- new_sat_writing_to_act_english(new_sat_vals)
new_act_vals <- data.frame(new_sat_vals, act_vals)
print(new_act_vals)

new_act_vals \%>\%
    mutate(act_vals_piped = new_sat_to_act(new_sat_vals))
}
