% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sat_writing_to_old_sat_writing.R
\name{new_sat_writing_to_old_sat_writing}
\alias{new_sat_writing_to_old_sat_writing}
\title{Recode new SAT writing and language test to old SAT writing section}
\usage{
new_sat_writing_to_old_sat_writing(x)
}
\arguments{
\item{x}{numeric vector of new SAT writing and language values}
}
\value{
integer vector of old SAT writing section values
}
\description{
Function to recode new SAT writing and language test to old SAT writing section
}
\examples{
require(dplyr)

new_sat_vals <- sample(10:40, 10)
old_sat_vals <- new_sat_writing_to_old_sat_writing(new_sat_vals)
new_old_vals <- data.frame(new_sat_vals, old_sat_vals)
print(new_old_vals)

new_old_vals \%>\%
    mutate(old_sat_vals_piped = new_sat_writing_to_old_sat_writing(new_sat_vals))
}
