% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_math_to_new_sat_math_section.R
\name{old_sat_math_to_new_sat_math_section}
\alias{old_sat_math_to_new_sat_math_section}
\title{Recode old SAT math section to new SAT math section (200-800 scale)}
\usage{
old_sat_math_to_new_sat_math_section(x)
}
\arguments{
\item{x}{numeric vector of old SAT math section values}
}
\value{
integer vector of new SAT math section values (200-800 scale)
}
\description{
Function to recode old SAT math section to new SAT math section
(200-800 scale)
}
