% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_reading_to_new_sat_reading.R
\name{old_sat_reading_to_new_sat_reading}
\alias{old_sat_reading_to_new_sat_reading}
\title{Recode old SAT critical reading test to new SAT reading section}
\usage{
old_sat_reading_to_new_sat_reading(x)
}
\arguments{
\item{x}{numeric vector of old SAT critical reading test values}
}
\value{
integer vector of new SAT reading section values
}
\description{
Function to recode old SAT critical reading test to new SAT reading section
}
