% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_to_new_sat_1600.R
\name{old_sat_to_new_sat_1600}
\alias{old_sat_to_new_sat_1600}
\title{Recode old SAT to new SAT (1600 scale)}
\usage{
old_sat_to_new_sat_1600(x)
}
\arguments{
\item{x}{numeric vector of old SAT values on 1600 scale}
}
\value{
integer vector of new SAT values on 1600 scale
}
\description{
Function to recode old SAT to new SAT (1600 scale)
}
