% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_writing_to_new_sat_writing.R
\name{old_sat_writing_to_new_sat_writing}
\alias{old_sat_writing_to_new_sat_writing}
\title{Recode old SAT writing test to new SAT writing and language  section}
\usage{
old_sat_writing_to_new_sat_writing(x)
}
\arguments{
\item{x}{numeric vector of old SAT writing values}
}
\value{
integer vector of new SAT writing and language section values
}
\description{
Function to recode old SAT writing test to new SAT writing and language section
}
\examples{
require(dplyr)

old_sat_vals <- round(sample(200:800, 10), -1)
new_sat_vals <- old_sat_writing_to_new_sat_writing(old_sat_vals)
old_new_vals <- data.frame(old_sat_vals, new_sat_vals)
print(old_new_vals)

old_new_vals \%>\%
    mutate(new_sat_vals_piped = old_sat_writing_to_new_sat_writing(old_sat_vals))
}
