% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sat_writingreading_to_new_ebrw.R
\name{old_sat_writingreading_to_new_ebrw}
\alias{old_sat_writingreading_to_new_ebrw}
\title{Recode old SAT writing plus critical reading sections to new SAT EBRW section}
\usage{
old_sat_writingreading_to_new_ebrw(x)
}
\arguments{
\item{x}{numeric vector of old SAT writing plus critical reading section values}
}
\value{
integer vector of new SAT EBRW section values
}
\description{
Function to recode old SAT writing plus critical reading sections
to new SAT Evidence-Based Reading and Writing Section test
}
